% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_seasons.R
\name{append_seasons}
\alias{append_seasons}
\title{append_seasons}
\usage{
append_seasons(x)
}
\arguments{
\item{x}{A \code{data.frame} of fixtures with a \code{date} column}
}
\value{
This function returns a \code{data.frame} with the following column appended:
\itemize{
 \item{season}{Character values representing the season that each match was played in.}
}
}
\description{
Using the date of the fixture, the season is appended to a data frame of fixtures.
Tyopically seasons run from August to May however Covid rescheduling increased the length of the season in 2021.
}
\examples{
\dontrun{
df <- append_seasons()
}
}
\author{
Daf Howells <dafhowells@yahoo.co.uk>
}
