% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prem_form.R
\name{get_prem_form}
\alias{get_prem_form}
\title{get_prem_form
Calculate rolling averages for points accumulated and goals scored for each team home and away.
This function creates a base level data frame for model fitting.
It returns a breakdown of each team's average points accumulated, goals scored and goals conceded home and away.
From this we can assess a team's relative attacking and defending strengths to use as variables in a model.}
\usage{
get_prem_form(
  prem_history = get_prem_history(),
  home_form_games_lag = 4,
  home_scored_games_lag = 4,
  home_conceded_games_lag = 4,
  away_form_games_lag = 4,
  away_scored_games_lag = 4,
  away_conceded_games_lag = 4
)
}
\arguments{
\item{prem_history}{a \code{data.frame} of past fixtures given by \code{get_prem_history()}}

\item{home_form_games_lag}{The count of lagged games to calculate average home points. Default is 4}

\item{home_scored_games_lag}{The count of lagged games to calculate average home goals scored. Default is 4,}

\item{home_conceded_games_lag}{The count of lagged games to calculate average home goals conceded. Default is 4,}

\item{away_form_games_lag}{The count of lagged games to calculate average away points. Default is 4,}

\item{away_scored_games_lag}{The count of lagged games to calculate average away goals scored. Default is 4,}

\item{away_conceded_games_lag}{The count of lagged games to calculate average away goals conceded Default is 4,}
}
\value{
This function returns a \code{data.frame} with the following columns:
\itemize{
 \item{team}{Character representing a team in a fixture}
 \item{fixture}{Character represening the fixture i.e. Arsenal v Everton}
 \item{date}{Date when the fixture was played}
 \item{fthg}{Numeric fulltime home goals scored in the fixture}
 \item{ftag}{Numeric fulltime away goals scored in the fixture}ftag
 \item{form}{Numeric running average of points accumulated (3 for a win, 1 for a draw) either home or away}form
 \item{scored}{Numeric running average of goals scored either home or away}
 \item{conceded}{Numeric running average of goals conceded either home or away}
 \item{home}{Numeric logical represnting whether the team was home or away}
 \item{win}{Numeric logical represnting whether the team won the fixture}
 \item{draw}{Numeric logical represnting whether the team drew the fixture}
}
}
\description{
get_prem_form
Calculate rolling averages for points accumulated and goals scored for each team home and away.
This function creates a base level data frame for model fitting.
It returns a breakdown of each team's average points accumulated, goals scored and goals conceded home and away.
From this we can assess a team's relative attacking and defending strengths to use as variables in a model.
}
\examples{
\dontrun{
prem_history <- get_prem_history()
prem_form_df <- get_prem_form(prem_history)
}
}
\author{
Daf Howells <dafhowells@yahoo.co.uk>
}
