% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_xg_fixture_history.R
\name{get_xg_fixture_history}
\alias{get_xg_fixture_history}
\title{get_xg_fixture_history
Calculate rolling averages expected goals (xG) scored for each team home and away for the same fixture over time.
This function scrapes fixture level xG data from https://fbref.com.
It returns a breakdown of each team's rolling average xG for and against both home and away.
From this we can assess a team's relative attacking and defending strengths to use as variables in a model.}
\usage{
get_xg_fixture_history(
  home_scored_games_lag = 4,
  home_conceded_games_lag = 4,
  away_scored_games_lag = 4,
  away_conceded_games_lag = 4
)
}
\arguments{
\item{home_scored_games_lag}{The count of lagged games to calculate average home xG. Default is 4,}

\item{home_conceded_games_lag}{The count of lagged games to calculate average home xG against. Default is 4,}

\item{away_scored_games_lag}{The count of lagged games to calculate average away xG. Default is 4,}

\item{away_conceded_games_lag}{The count of lagged games to calculate average away xG against. Default is 4,}
}
\value{
This function returns a data.frame with the following columns
}
\description{
get_xg_fixture_history
Calculate rolling averages expected goals (xG) scored for each team home and away for the same fixture over time.
This function scrapes fixture level xG data from https://fbref.com.
It returns a breakdown of each team's rolling average xG for and against both home and away.
From this we can assess a team's relative attacking and defending strengths to use as variables in a model.
}
\examples{
\dontrun{
xg_fixture_history_df <- get_xg_fixture_history()
}
}
\author{
Daf Howells <dafhowells@yahoo.co.uk>
}
