% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prem_2020.R
\docType{data}
\name{prem_2020}
\alias{prem_2020}
\title{Premier League Results for the 2020 season}
\format{
A data frame of 380 rows and 6 columns
\describe{
 \item{div}{Character values representing the division each match was played in. Default to E0 for the Premier League}
 \item{date}{Character values for the date that the match was played}
 \item{home_team}{Character values for the home team in the fixture}
 \item{away_team}{Character values for the away team in the fixture}
 \item{fthg}{Character values for the number of goals scored by the home team in the fixture}
 \item{ftag}{Character values for the number of goals scored by the away team in the fixture}
}
}
\source{
https://datahub.io/sports-data/english-premier-league
}
\usage{
prem_2020
}
\description{
Premier League football match results from the 2020 season
}
\keyword{datasets}
