% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prem_pre_2018.R
\docType{data}
\name{prem_pre_2018}
\alias{prem_pre_2018}
\title{Premier League Results pre 2018}
\format{
A data frame of 9664 rows and 6 columns
\describe{
 \item{div}{Character values representing the division each match was played in. Default to E0 for the Premier League}
 \item{date}{Character values for the date that the match was played}
 \item{home_team}{Character values for the home team in the fixture}
 \item{away_team}{Character values for the away team in the fixture}
 \item{fthg}{Character values for the number of goals scored by the home team in the fixture}
 \item{ftag}{Character values for the number of goals scored by the away team in the fixture}
}
}
\source{
https://datahub.io/sports-data/english-premier-league
}
\usage{
prem_pre_2018
}
\description{
Premier League football match results from before 2018
}
\keyword{datasets}
