% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randy.R
\docType{package}
\name{randy}
\alias{randy}
\alias{randy-package}
\title{randy: Data and Functions for Modelling Premier League Goals}
\description{
Source historic premier league data to be used for modelling and predicting the number of goals that each team will score.
}
