% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prem_history.R
\name{get_prem_history}
\alias{get_prem_history}
\title{get_prem_history
Extract current season Premier League match data and combine it with previous seasons..
Current season sourced from Fantasy League API.}
\usage{
get_prem_history()
}
\value{
This function returns a \code{data.frame} with the following columns:
\itemize{
 \item{div}{Character representing the division, default E0 which is the value for the Premier League}
 \item{date}{Date when the fixture was played}
 \item{home_team}{Character showing team playing at home}
 \item{home_team}{Character showing team playing away}
 \item{fthg}{Numeric fulltime home goals scored in the fixture}
 \item{ftag}{Numeric fulltime away goals scored in the fixture}
}
}
\description{
get_prem_history
Extract current season Premier League match data and combine it with previous seasons..
Current season sourced from Fantasy League API.
}
\examples{
\dontrun{
prem_history <- get_prem_history()
}
}
\author{
Daf Howells <dafhowells@yahoo.co.uk>
}
