% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_xg_fixture_history.R
\name{get_xg_fixture_history}
\alias{get_xg_fixture_history}
\title{Calculate rolling averages expected goals (xG) scored for each team home and away}
\usage{
get_xg_fixture_history(
  home_scored_games_lag = 4,
  home_conceded_games_lag = 4,
  away_scored_games_lag = 4,
  away_conceded_games_lag = 4
)
}
\arguments{
\item{home_scored_games_lag}{The count of lagged games to calculate average home xG. Default is 4,}

\item{home_conceded_games_lag}{The count of lagged games to calculate average home xG against. Default is 4,}

\item{away_scored_games_lag}{The count of lagged games to calculate average away xG. Default is 4,}

\item{away_conceded_games_lag}{The count of lagged games to calculate average away xG against. Default is 4,}
}
\value{
This function returns a data.frame with the following columns
\describe{
 \item{date}{Date when the fixture was played}
 \item{team}{Character representing a team in a fixture}
 \item{fixture}{Character represening the fixture e.g Arsenal v Everton}
 \item{actual_score}{Character represening the actual score in the fixture e.g. 2-1}
 \item{xg1}{Numeric fulltime home xG in the fixture}
 \item{xg2}{Numeric fulltime away xG in the fixture}
 \item{home}{Logical represening whether or not the team was home or away in the fixture}
 \item{xg_for}{Numeric showing running average of xG for the team}
 \item{xg_against}{umeric showing running average of xG against the team}
}
}
\description{
This function returns a breakdown of each team's rolling average xG for and against both home and away.
}
\examples{
\dontrun{
xg_fixture_history_df <- get_xg_fixture_history()
}
}
\author{
Daf Howells <dafhowells@yahoo.co.uk>
}
